refactor: memory-usage tests should use a temp directory
The test for memory consumption creates a large log file.
Previously, it was creating it in the local test directory
which clutters up the project dir and could potentially
get committed. This change restructures the test to use
a temp directory as given by the OS.

Semver: patch
